Upgrade Elwood adapter from v2 to v3
