+++ .changeset/pink-penguins-applaud.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/elwood-adapter': major
+---
+
+Upgraded EA framework from v2 to v3
