+++ packages/core/legos/tsconfig.json
@@ -188,9 +188,6 @@
       "path": "../../sources/elwood"
     },
     {
-      "path": "../../sources/elwood-test"
-    },
-    {
       "path": "../../sources/ens"
     },
     {
