--- packages/sources/elwood-test/CHANGELOG.md
@@ -1,44 +1,0 @@
-# @chainlink/elwood-test-adapter
-
-## 1.1.2
-
-### Patch Changes
-
-- b98296089: Bumped v3 framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 255af9e8a: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d3b95f80e: Added error caching for failed subscribe/unsubscribe messages
-
-## 1.0.3
-
-### Patch Changes
-
-- f850f4831: Bump framework version
-- 917400c7b: Bump framework version
-
-## 1.0.2
-
-### Patch Changes
-
-- 1542a53c2: Bump framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- d369bce53: Fix name
-
-## 1.0.0
-
-### Major Changes
-
-- 919ab1ca3: Initial release
