--- packages/sources/elwood-test/src/config/index.ts
@@ -1,18 +1,0 @@
-export const customSettings = {
-  API_KEY: {
-    description: 'API key',
-    type: 'string',
-    required: true,
-    sensitive: true,
-  },
-  WS_API_ENDPOINT: {
-    description: 'The websocket url for coinmetrics',
-    type: 'string',
-    default: 'wss://api.chk.elwood.systems/v1/stream',
-  },
-  API_ENDPOINT: {
-    description: 'The API url for coinmetrics',
-    type: 'string',
-    default: 'https://api.chk.elwood.systems/v1/stream',
-  },
-} as const
