--- packages/sources/elwood-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { cryptoEndpoint } from './crypto'
