--- packages/sources/elwood-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "from": "LINK",
-  "to": "USD"
- }]
-}
