+++ packages/sources/elwood/CHANGELOG.md
@@ -1,111 +1,44 @@
 # @chainlink/elwood-adapter
 
-## 1.0.11
+## 1.1.2
 
 ### Patch Changes
 
-- Updated dependencies [fc46b78fc]
-- Updated dependencies [1de0689c6]
-  - @chainlink/ea-bootstrap@2.25.0
-  - @chainlink/ea-test-helpers@1.4.1
+- b98296089: Bumped v3 framework version
 
-## 1.0.10
+## 1.1.1
 
 ### Patch Changes
 
-- Updated dependencies [e8576df4e]
-- Updated dependencies [842651f93]
-  - @chainlink/ea-bootstrap@2.24.0
-  - @chainlink/ea-test-helpers@1.4.1
+- 255af9e8a: Bumped v3 framework version
 
-## 1.0.9
+## 1.1.0
 
-### Patch Changes
+### Minor Changes
 
-- Updated dependencies [13eb04f5a]
-- Updated dependencies [221ab1e5f]
-  - @chainlink/ea-bootstrap@2.23.0
-  - @chainlink/ea-test-helpers@1.4.1
-
-## 1.0.8
-
-### Patch Changes
-
-- Updated dependencies [572b89314]
-- Updated dependencies [068dd3672]
-  - @chainlink/ea-bootstrap@2.22.2
-  - @chainlink/ea-test-helpers@1.4.1
-
-## 1.0.7
-
-### Patch Changes
-
-- Updated dependencies [26b046b1e]
-  - @chainlink/ea-bootstrap@2.22.1
-  - @chainlink/ea-test-helpers@1.4.1
-
-## 1.0.6
-
-### Patch Changes
-
-- Updated dependencies [b8061e1d5]
-  - @chainlink/ea-bootstrap@2.22.0
-  - @chainlink/ea-test-helpers@1.4.1
-
-## 1.0.5
-
-### Patch Changes
-
-- Updated dependencies [3c1a320b5]
-  - @chainlink/ea-bootstrap@2.21.0
-  - @chainlink/ea-test-helpers@1.4.1
-
-## 1.0.4
-
-### Patch Changes
-
-- Updated dependencies [b9982adc8]
-- Updated dependencies [f710272c6]
-- Updated dependencies [991fc76af]
-  - @chainlink/ea-bootstrap@2.20.0
-  - @chainlink/ea-test-helpers@1.4.1
+- d3b95f80e: Added error caching for failed subscribe/unsubscribe messages
 
 ## 1.0.3
 
 ### Patch Changes
 
-- Updated dependencies [5e7393deb]
-- Updated dependencies [5e7393deb]
-  - @chainlink/ea-bootstrap@2.19.3
-  - @chainlink/ea-test-helpers@1.4.1
+- f850f4831: Bump framework version
+- 917400c7b: Bump framework version
 
 ## 1.0.2
 
 ### Patch Changes
 
-- Updated dependencies [3a0e5aaa9]
-  - @chainlink/ea-bootstrap@2.19.2
-  - @chainlink/ea-test-helpers@1.4.1
+- 1542a53c2: Bump framework version
 
 ## 1.0.1
 
 ### Patch Changes
 
-- Updated dependencies [05a3f9464]
-  - @chainlink/ea-bootstrap@2.19.1
-  - @chainlink/ea-test-helpers@1.4.1
+- d369bce53: Fix name
 
 ## 1.0.0
 
 ### Major Changes
 
-- 530753225: Add new Elwood EA
-
-### Patch Changes
-
-- Updated dependencies [5a1adab07]
-- Updated dependencies [ed54a688b]
-- Updated dependencies [530753225]
-- Updated dependencies [88fdcb137]
-  - @chainlink/ea-bootstrap@2.19.0
-  - @chainlink/ea-test-helpers@1.4.1
+- 919ab1ca3: Initial release
