+++ packages/sources/elwood/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@chainlink/elwood-adapter",
-  "version": "1.0.11",
+  "version": "1.1.2",
   "description": "Chainlink Elwood adapter.",
   "keywords": [
     "Chainlink",
@@ -28,16 +28,17 @@
     "start": "yarn server:dist"
   },
   "dependencies": {
-    "@chainlink/ea-bootstrap": "workspace:*",
-    "@chainlink/ea-test-helpers": "workspace:*",
+    "@chainlink/external-adapter-framework": "0.14.5",
+    "axios": "0.27.2",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@types/jest": "27.0.3",
-    "@types/node": "16.11.41",
+    "@types/jest": "27.5.2",
+    "@types/node": "16.11.51",
     "@types/supertest": "2.0.12",
-    "nock": "13.2.7",
-    "supertest": "6.2.3",
+    "mock-socket": "^9.1.5",
+    "nock": "13.2.9",
+    "supertest": "6.2.4",
     "typescript": "4.7.4"
   }
 }
