--- packages/sources/elwood/schemas/env.json
@@ -1,18 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/elwood-adapter.json",
-  "title": "Elwood Adapter",
-  "description": "This adapter implements a WS interface for price data from the [Elwood](https://elwood.io/) platform. Note that the first request for a price will fail because the WS connection does not have enough time to start up before the first subscription request is sent.",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "API key to use (combined WS and HTTP key)"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
