--- packages/sources/elwood/src/config/includes.json
@@ -1,1 +1,0 @@
-{}
