--- packages/sources/elwood/src/config/limits.json
@@ -1,4 +1,0 @@
-{
-  "http": {},
-  "ws": {}
-}
