--- packages/sources/elwood/src/config/symbols.json
@@ -1,1 +1,0 @@
-{}
