+++ packages/sources/elwood/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as PriceInputParameters } from './price'
-
-export type TInputParameters = PriceInputParameters
-
-export * as price from './price'
+export { cryptoEndpoint } from './crypto'
