--- packages/sources/elwood/src/types.ts
@@ -1,24 +1,0 @@
-export type SubscribeRequest = {
-  action: 'subscribe' | 'unsubscribe'
-  stream: 'index'
-  symbol: string
-  index_freq: number
-}
-
-export type PriceResponse = {
-  type: 'Index'
-  data: {
-    price: string
-    symbol: string
-    timestamp: string
-  }
-  sequence: number
-}
-
-export type HeartbeatResponse = {
-  type: 'heartbeat'
-  data: string
-  sequence: number
-}
-
-export type ResponseMessage = PriceResponse | HeartbeatResponse
