+++ packages/sources/elwood/test/integration/adapter.test.ts
@@ -1,68 +1,60 @@
-import { AdapterRequest, FastifyInstance } from '@chainlink/ea-bootstrap'
+import { ServerInstance } from '@chainlink/external-adapter-framework'
+import { WebSocketClassProvider } from '@chainlink/external-adapter-framework/transports'
 import { AddressInfo } from 'net'
-import request, { SuperTest, Test } from 'supertest'
-import { server as startServer } from '../../src'
-import {
-  mockSubscribeWSResponse,
-  mockUnsubscribeWSResponse,
-  mockSubscribeResponse,
-  mockUnsubscribeResponse,
-} from './fixtures'
-import {
-  mockWebSocketFlow,
-  mockWebSocketProvider,
-  mockWebSocketServer,
-  MockWsServer,
-} from '@chainlink/ea-test-helpers'
-import { WebSocketClassProvider } from '@chainlink/ea-bootstrap/dist/lib/middleware/ws/recorder'
 import * as nock from 'nock'
+import request, { SuperTest, Test } from 'supertest'
+import { mockSubscribeError, mockSubscribeResponse, mockUnsubscribeResponse } from './fixtures'
+import { mockWebSocketProvider, mockWebSocketServer } from './setup'
 
 describe('websocket', () => {
-  let mockedWsServer: InstanceType<typeof MockWsServer>
-  let fastify: FastifyInstance
+  let fastify: ServerInstance
   let req: SuperTest<Test>
   let oldEnv: NodeJS.ProcessEnv
+  let spy: jest.SpyInstance
+  const apiKey: string = process.env['API_KEY'] ?? 'test-api-key'
 
   beforeAll(async () => {
     oldEnv = JSON.parse(JSON.stringify(process.env))
-    process.env.WS_ENABLED = 'true'
-    process.env.API_KEY = process.env.API_KEY ?? 'test-api-key'
-    process.env.WS_SUBSCRIPTION_TTL = '3000'
+    process.env['API_KEY'] = apiKey
+    process.env['WS_SUBSCRIPTION_TTL'] = '3000'
+    process.env['METRICS_ENABLED'] = 'false'
+    process.env['RATE_LIMIT_CAPACITY_SECOND'] = '10'
 
-    if (!process.env.RECORD) {
-      mockedWsServer = mockWebSocketServer('wss://api.chk.elwood.systems/v1/stream')
+    const mockDate = new Date('2022-05-10T16:09:27.193Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
+
+    if (!process.env['RECORD']) {
+      mockWebSocketServer(`wss://api.chk.elwood.systems/v1/stream?apiKey${process.env['API_KEY']}`)
       mockWebSocketProvider(WebSocketClassProvider)
     } else {
       nock.recorder.rec()
     }
 
-    fastify = await startServer()
+    const { server } = await import('../../src')
+    const api = await server()
+    if (!api) {
+      throw 'API did not start'
+    }
+    fastify = api
     req = request(`localhost:${(fastify.server.address() as AddressInfo).port}`)
   })
 
   afterAll((done) => {
+    spy.mockRestore()
     process.env = oldEnv
     fastify.close(done)
   })
 
   describe('price endpoint', () => {
     it('should return success', async () => {
-      mockSubscribeResponse()
-      mockUnsubscribeResponse()
-      const data: AdapterRequest = {
+      mockSubscribeResponse(apiKey)
+      mockUnsubscribeResponse(apiKey)
+      const data = {
         id: '1',
         data: {
           base: 'ETH',
           quote: 'USD',
         },
-      }
-
-      let flowFulfilled: Promise<boolean> | undefined
-      if (!process.env.RECORD) {
-        flowFulfilled = mockWebSocketFlow(mockedWsServer, [
-          mockSubscribeWSResponse,
-          mockUnsubscribeWSResponse,
-        ])
       }
 
       const response = await req
@@ -74,7 +66,25 @@ describe('websocket', () => {
         .expect(200)
 
       expect(response.body).toMatchSnapshot()
-      return !process.env.RECORD ? await flowFulfilled : undefined
+    }, 10_000)
+    it('should return cached subscribe error', async () => {
+      mockSubscribeError(apiKey)
+      const data = {
+        data: {
+          base: 'XXX',
+          quote: 'USD',
+        },
+      }
+
+      const response = await req
+        .post('/')
+        .send(data)
+        .set('Accept', '*/*')
+        .set('Content-Type', 'application/json')
+        .expect('Content-Type', /json/)
+        .expect(400)
+
+      expect(response.body).toMatchSnapshot()
     }, 10_000)
   })
 })
