+++ packages/sources/elwood/test/integration/fixtures.ts
@@ -1,9 +1,9 @@
 import nock from 'nock'
 
-export const mockSubscribeResponse = () => {
-  nock('https://api.chk.elwood.systems:443', { encodedQueryParams: true })
+export const mockSubscribeResponse = (apiKey: string) => {
+  nock(`https://api.chk.elwood.systems`, { encodedQueryParams: true })
     .persist()
-    .post('/v1/stream', {
+    .post(`/v1/stream?apiKey=${apiKey}`, {
       action: 'subscribe',
       stream: 'index',
       symbol: 'ETH-USD',
@@ -12,10 +12,10 @@ export const mockSubscribeResponse = () => {
     .reply(200, {}, [])
 }
 
-export const mockUnsubscribeResponse = () => {
-  nock('https://api.chk.elwood.systems:443', { encodedQueryParams: true })
+export const mockUnsubscribeResponse = (apiKey: string) => {
+  nock(`https://api.chk.elwood.systems`, { encodedQueryParams: true })
     .persist()
-    .post('/v1/stream', {
+    .post(`/v1/stream?apiKey=${apiKey}`, {
       action: 'unsubscribe',
       stream: 'index',
       symbol: 'ETH-USD',
@@ -24,19 +24,24 @@ export const mockUnsubscribeResponse = () => {
     .reply(200, {}, [])
 }
 
-export const mockSubscribeWSResponse = {
-  request: { action: 'subscribe', stream: 'index', symbol: 'ETH-USD', index_freq: 1000 },
-  response: [
-    { type: 'heartbeat', data: '2022-08-22T10:58:30.537993108Z', sequence: 1 },
-    {
-      type: 'Index',
-      data: { price: '10000.000', symbol: 'ETH-USD', timestamp: '2022-08-22T10:58:34.31724727Z' },
-      sequence: 2,
-    },
-  ],
-}
-
-export const mockUnsubscribeWSResponse = {
-  request: { action: 'unsubscribe', stream: 'index', symbol: 'ETH-USD', index_freq: 1000 },
-  response: [{ type: 'heartbeat', data: '2022-08-22T10:58:30.537993108Z', sequence: 3 }],
+export const mockSubscribeError = (apiKey: string) => {
+  nock(`https://api.chk.elwood.systems`, { encodedQueryParams: true })
+    .persist()
+    .post(`/v1/stream?apiKey=${apiKey}`, {
+      action: 'subscribe',
+      stream: 'index',
+      symbol: 'XXX-USD',
+      index_freq: 1000,
+    })
+    .reply(400, {
+      error: {
+        code: 400,
+        message: 'Symbol is not supported',
+        errors: [
+          {
+            domain: 'stream',
+          },
+        ],
+      },
+    })
 }
