+++ packages/tsconfig.json
@@ -300,9 +300,6 @@
       "path": "./sources/elwood"
     },
     {
-      "path": "./sources/elwood-test"
-    },
-    {
       "path": "./sources/ens"
     },
     {
