+++ packages/tsconfig.test.json
@@ -300,9 +300,6 @@
       "path": "./sources/elwood/tsconfig.test.json"
     },
     {
-      "path": "./sources/elwood-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/ens/tsconfig.test.json"
     },
     {
